requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
fonttools==4.56.0
frozenlist==1.5.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.0
multidict==6.1.0
numpy==2.2.3
packaging==24.2
pillow==11.1.0
propcache==0.2.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
setuptools==75.8.0
six==1.17.0
yarl==1.18.3
